pyproject.toml
[project]
name = "berich"
version = "0.1.0"
description = "BeRich - Real-time trading bot for KRX and US markets using KIS API"
requires-python = ">=3.12"
dependencies = [
    # Async
    "aiohttp>=3.9",
    "websockets>=12.0",

    # Data processing
    "pandas>=2.0",
    "numpy>=1.24",
    "pandas-ta>=0.4.71b0",

    # Database
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.19",
    "aiomysql>=0.2",
    "cryptography>=42.0",

    # Config
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",

    # Utils
    "python-dateutil>=2.8",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "mypy>=1.8",
    "ruff>=0.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
